prometheus: fix etcd client cert being re-created on every terraform apply
I don't fully understand this, but I found that at the moment when
terraform is calculating the plan for the tls_locally_signed_cert,
the proposed new values of ca_cert_pem and ca_private_key_pem are
both the empty string, even though the correct value seems to get
filled in later.

I noticed that we get "Reading..." / "Reading complete." logs for
the top-level data.terraform_remote_state.talos but not the one
in the prometheus module. I suspect there is some weirdness with
data sources that have the same address within different modules.

Refactoring to use a single data.terraform_remote_state.talos at
the top level and passing the data across the module boundary seems
to resolve the problem.