invalidate Stamp version pins in flake.lock files
The Stamp version is pinned in modules/stamp_image/main.tf.
The flake.lock files need to be syntactically valid, but the version
pin for Stamp always gets overriden on the nix command line (in
Stamp's derivation.sh).

To save having to maintain a version pin in several places, and to
avoid accidentally building images with an unintended version of
Stamp, let's make it so you _have_ to use --override-input for the
flakes to be usable.